backend/requirements.txt
# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Google Agent Dev Kit
google-adk

# File upload handling
python-multipart>=0.0.6

# Docker integration
docker>=6.1.3

# Utilities
python-dotenv>=1.0.0

# Video processing (for local development)
opencv-python>=4.8.0
pillow>=10.0.0
moviepy>=1.0.0

# Development
pytest>=7.4.3

# OpenAI
openai